Rename package
This should match the directory and lambda name
