+++ scala/lambdas/preingest-tdr-aggregator/src/main/scala/uk/gov/nationalarchives/preingesttdraggregator/Lambda.scala
@@ -1,4 +1,4 @@
-package uk.gov.nationalarchives.eventaggregator
+package uk.gov.nationalarchives.preingesttdraggregator
 
 import cats.effect.unsafe.implicits.global
 import cats.effect.{IO, Ref}
@@ -7,13 +7,13 @@ import com.amazonaws.services.lambda.runtime.{Context, RequestHandler}
 import pureconfig.generic.derivation.default.*
 import pureconfig.module.catseffect.syntax.*
 import pureconfig.{ConfigCursor, ConfigReader, ConfigSource}
-import uk.gov.nationalarchives.eventaggregator.Lambda.*
-import uk.gov.nationalarchives.eventaggregator.Aggregator.*
+import uk.gov.nationalarchives.preingesttdraggregator.Lambda.*
+import uk.gov.nationalarchives.preingesttdraggregator.Aggregator.*
 import uk.gov.nationalarchives.{DADynamoDBClient, DASFNClient}
 import io.circe.generic.auto.*
 import pureconfig.ConfigReader.Result
-import uk.gov.nationalarchives.eventaggregator.Duration.Seconds
-import uk.gov.nationalarchives.eventaggregator.Ids.GroupId
+import uk.gov.nationalarchives.preingesttdraggregator.Duration.Seconds
+import uk.gov.nationalarchives.preingesttdraggregator.Ids.GroupId
 
 import java.net.URI
 import java.time.Instant
